requirements.txt
Flask==3.0.3
gunicorn==23.0.0
werkzeug==3.0.3

# Image processing
numpy==1.26.4
Pillow==10.4.0
scikit-image==0.24.0
opencv-python-headless==4.10.0.84

# ML
tensorflow==2.16.2
keras==3.11.3
